pyproject.toml
[tool.poetry]
name = "qshed-client"
version = "0.1.1"
description = ""
authors = ["nixoninnes"]
packages = [
    { include = "qshed"}
]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.4.2"
requests = "^2.27.1"
pydantic = "^1.9.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
